Feature: Added Logout functionality
