--- app/src/main/java/com/tip/lunchbox/view/fragment/FoodFragment.java
@@ -1,25 +1,0 @@
-package com.tip.lunchbox.view.fragment;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import androidx.fragment.app.Fragment;
-
-import com.tip.lunchbox.R;
-
-public class FoodFragment extends Fragment {
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater,
-                              ViewGroup container,
-                              Bundle savedInstanceState) {
-
-        return inflater.inflate(R.layout.fragment_food,container,false);
-    }
-}
